Bring all testings in gitlab and travis CI to Azure Pipelines
This expands current Azure Pipelines Windows host tools build
testing to cover all the CI testing in gitlab and travis CI.

Note for some unknown reason, the 'container' cannot be used for
any jobs that have buildman, for buildman does not exit properly
and hangs the job forever. As a workaround, we manually call
docker to run the image to perform the CI tasks.

A complete run on Azure Pipelines takes about 2 hours and 10
minutes.

Signed-off-by: Bin Meng <bmeng.cn@gmail.com>

Series-version: 2
Series-to: Tom, U-Boot
Cover-letter:
Bring all testings in gitlab and travis CI to Azure Pipelines
At present we have set up a Microsoft Azure Pipelines to build
U-Boot host tools for Windows. We can expand it to cover all the
CI testing in gitlab and travis CI.

There are issues if we use the 'container' keyword in the pipeline,
that buildman seems to not able to exit correctly and hangs forever.
So we manually call docker to run the image to perform the CI tasks.
Unfortunately with above configuration, the uboot user inside the
container does not have the write permission to the U-Boot git repo,
and this exposes several issues in the U-Boot build process.

With all such issues fixed, a complete run on Azure Pipelines takes
about 2 hours and 10 minutes. It seems to be a good backup for the
free travis CI service.

This series depends on
http://patchwork.ozlabs.org/project/uboot/list/?series=138998

and is avaiable at u-boot-x86/azure for testing.
END

Commit-notes:
See the build result at:
https://dev.azure.com/bmeng/GitHub/_build/results?buildId=109
END