x86: qemu: Enable ramfb by default
Now that we have everything in place to support ramfb, let's wire it up
by default in the x86 QEMU targets. That way, we can use ramfb graphical
console instead of the default by passing -vga none -device ramfb to the
QEMU command line.

Also increase SYS_MALLOC_F_LEN for QEMU x86_64 to be the same as its SPL
counterpart, because we're running out of alloc space in pre-reloc stage
with ramfb enabled.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>
[Bin: Remove <common.h>, and rebase on u-boot/master]
Signed-off-by: Bin Meng <bmeng.cn@gmail.com>